fake-vm: Avoid panic'ing in the deadline handler
panic() is not well handled inside a signal handler and can corrupt
the stack above the signal handler frame. This is even worse with cargo
test since further tests may fail/freeze/report corrupt state in an
unpredictable manner.

For the moment, replace .expect() with error catching and terminate the
simulation protocol, although the application thread may keep on
running.

A better solution would consist in catching panic() at the top of the
signal handler.