vsg: Handle src/dest port in mesages
src and destination are now filled in a packet.
They store (ipv4+port) information so that:
- the coordinator can read the dest to route the message
- the application can read the port to dispatch to the right endpoint

Limitations: this makes explicit the use of src and port (they could be
embeded in the payload). This also makes vsg rely on something like
udp/tcp. Let's say this is good enough for now.

Additional change:
- add a more testable example (constant_rate)
- some other stuffs (I promise I'll try to get more atomic commits in
the future :())