boot.py: now use a more persistent working dir
When a VM start a working dir is created prefixed by the hostname of the
VM.
It'll contains all the file needed to boot the VM (cloud-init, base
qcow image).
Also it can be used to store specific files (pcap dumped file ...)
coming from the extra args passed to qemu.

Note the stdout file is store there by default (`out`)