tests: Fix exit behavior of the fake actor
Once the fake actor scenario is over, the fake actor starts running
(again) the client-side of the test. Although this is usually harmless
because the client-side quickly blocks waiting for the server to
accept the connection to its socket, fix this by exiting after making
sure that the client side is done (either by destroying ScenarioRunner
or by exiting abnormally).