tansiv.py: align wrapper requirement
when running a TS, tansiv injects the socket name as the first argument
of the script that is fork/exec-ed.
notansiv.py is the symetric to tansiv but for nts system but use the
same wrapper script at the end (boot.py). So we make notansiv also
inject a (dummy) socket name as the first parameter of the boot.py
script.